fs: deduplicate placeholders for generic and xattrs
