+++ internal/fs/node_noxattr.go
@@ -1,0 +1,18 @@
+//go:build aix || netbsd || openbsd
+// +build aix netbsd openbsd
+
+package fs
+
+import (
+	"github.com/restic/restic/internal/restic"
+)
+
+// nodeRestoreExtendedAttributes is a no-op
+func nodeRestoreExtendedAttributes(_ *restic.Node, _ string) error {
+	return nil
+}
+
+// nodeFillExtendedAttributes is a no-op
+func nodeFillExtendedAttributes(_ *restic.Node, _ string, _ bool) error {
+	return nil
+}
